+++ changelog.md
@@ -1,3 +1,8 @@
+##Version 1.2.6.a (Build #136) -pt
+Added new pf_config::get('main_page') call.
+Added Users listing at index.php/users (with delete button)
+Added Entire User System Management System. Create/List/Delete users possible.
+
 ##Version 1.2.6.a (Build #135) -pt
 Added Config Server Page with default values set to whatever the server has now <br />
 Added Requirement of Sqlite 3 support in PHP (see readme.md)